fix: resolve React fast refresh warning in SentryErrorBoundary.tsx
- Move utility functions to new src/lib/sentryUtils.tsx file
- Keep only the SentryErrorBoundary component in the component file
- Enables proper React Fast Refresh for the component
