+++ src/components/SentryErrorBoundary.tsx
@@ -104,44 +104,3 @@ export class SentryErrorBoundary extends Component<Props, State> {
   }
 }
 
-/**
- * Higher-order component for adding error boundary protection to any component
- * 
- * @example
- * ```tsx
- * const ProtectedComponent = withSentryErrorBoundary(MyComponent, <div>Custom fallback</div>);
- * ```
- * 
- * @param Component The component to wrap with error boundary
- * @param fallback Optional custom fallback UI (defaults to built-in error UI)
- * @returns Component wrapped with SentryErrorBoundary
- */
-export function withSentryErrorBoundary<P extends object>(
-  Component: React.ComponentType<P>,
-  fallback?: ReactNode
-) {
-  return function WrappedComponent(props: P) {
-    return (
-      <SentryErrorBoundary fallback={fallback}>
-        <Component {...props} />
-      </SentryErrorBoundary>
-    );
-  };
-}
-
-/**
- * Hook-based error boundary using Sentry's built-in implementation
- * 
- * Use this for simpler integration when you don't need custom error UI:
- * 
- * @example
- * ```tsx
- * const MyComponent = SentryErrorBoundaryHook(MyBaseComponent, {
- *   fallback: ({ error }) => <div>Error: {error.message}</div>,
- *   beforeCapture: (scope, error) => {
- *     scope.setTag('section', 'user-profile');
- *   }
- * });
- * ```
- */
-export const SentryErrorBoundaryHook = Sentry.withErrorBoundary;
